requirements.txt
requests~=2.31.0
loguru~=0.7.2
PyYAML~=6.0.1
watchdog~=4.0.0
python-telegram-bot~=21.0.1
